requirements.txt
pytest
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gast==0.2.2
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
GoogleAppEngineCloudStorageClient==1.9.22.1
grpcio==1.27.2
gunicorn==19.9.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
opt-einsum==3.2.1
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==3.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
